Tidy: Remove widget icons export from chat panel header, move to widgetIcons
