+++ app/utils/widgetIcons.tsx
@@ -5,6 +5,7 @@ import {
   ChartPolarIcon,
   ChartScatterIcon,
   ListNumbersIcon,
+  PresentationChartIcon,
   StackIcon,
 } from "@phosphor-icons/react";
 import type { Icon } from "@phosphor-icons/react";
@@ -13,9 +14,6 @@ import type { InsightWidget } from "@/app/types/chat";
 /**
  * Maps an InsightWidget type to the Phosphor icon *component* (not an
  * instance), so callers can pass their own size/color/weight props.
- *
- * Mirrors the `WidgetIcons` element map in ChatPanelHeader.tsx — kept in
- * sync manually for now since the two have different ergonomics.
  */
 export const WidgetIconComponent: Record<InsightWidget["type"], Icon> = {
   line: ChartLineIcon,
@@ -27,4 +25,18 @@ export const WidgetIconComponent: Record<InsightWidget["type"], Icon> = {
   scatter: ChartScatterIcon,
   table: ListNumbersIcon,
   "dataset-card": StackIcon,
+};
+
+/** Element-instance map for inline rendering without size/color overrides. */
+export const WidgetIcons = {
+  line: <ChartLineIcon />,
+  table: <ListNumbersIcon />,
+  bar: <ChartBarIcon />,
+  "stacked-bar": <ChartBarIcon />,
+  "grouped-bar": <ChartBarIcon />,
+  pie: <ChartPieSliceIcon />,
+  insight: <PresentationChartIcon />,
+  "dataset-card": <StackIcon />,
+  scatter: <ChartScatterIcon />,
+  area: <ChartPolarIcon />,
 };
